Cleanup enum scanner, and fix compiler warnings.
