+++ EasyCards.EnumScanner/EnumDefinition.cs
@@ -1,0 +1,2 @@
+#pragma warning disable SYSLIB1037
+public record EnumDefinition(string Name, string BaseType, bool Flags, EnumMemberDefinition[] Members);
