+++ EasyCards.EnumScanner/EnumDefinitions.cs
@@ -1,0 +1,2 @@
+#pragma warning disable SYSLIB1037
+public record EnumDefinitions(EnumDefinition[] Enums);
